Implement check for webhook presence in Push-PublicWebhookProcess.ps1 to prevent errors during entity removal, enhancing the function's robustness and error handling.

+++ Modules/CIPPCore/Public/Entrypoints/Activity Triggers/Webhooks/Push-PublicWebhookProcess.ps1
@@ -18,7 +18,9 @@ function Push-PublicWebhookProcess {
     } catch {
         Write-Host "Webhook Exception: $($_.Exception.Message)"
     } finally {
-        $Entity = $Webhook | Select-Object -Property RowKey, PartitionKey
-        Remove-AzDataTableEntity -Force @Table -Entity $Entity
+        if ($Webhook) {
+            $Entity = $Webhook | Select-Object -Property RowKey, PartitionKey
+            Remove-AzDataTableEntity -Force @Table -Entity $Entity
+        }
     }
 }
